# docs/swagger.yaml
basePath: /
definitions:
  auth.ChangePasswordRequest:
    properties:
      password:
        type: string
    type: object
  auth.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  view.LoginAuthView:
    properties:
      email:
        type: string
      email_verified_at:
        type: string
      id:
        type: string
      name:
        type: string
      refresh_token:
        type: string
      token:
        type: string
      token_expiration:
        type: integer
    type: object
...
  view.Response:
    properties:
      message:
        type: string
      payload: {}
      success:
        type: boolean
    type: object
  view.UserEmptyView:
    properties:
      id:
        type: string
    type: object
...
host: 127.0.0.1:8080
info:
  contact: {}
  description: Authentication API Service.
  title: Authentication Server
  version: "1.0"
paths:
  /auth/change-password/{token}:
    post:
      consumes:
      - application/json
      description: Verify forgot password
      parameters:
      - description: User email
        in: query
        name: email
        type: string
      - description: User token
        in: path
        name: token
        required: true
        type: string
      - description: User password
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/view.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/view.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/view.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/view.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/view.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/view.Response'
      tags:
      - auth
  /auth/email-resend:
    post:
      consumes:
      - application/json
      description: Resend email
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/view.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/view.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/view.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/view.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/view.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/view.Response'
      security:
      - JWT: []
      tags:
      - auth
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Forgot password
      parameters:
      - description: User email
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/view.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/view.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/view.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/view.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/view.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/view.Response'
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user
      parameters:
      - description: User email and password
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/view.Response'
            - properties:
                payload:
                  $ref: '#/definitions/view.LoginAuthView'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/view.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/view.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/view.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/view.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/view.Response'
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/view.Response'
            - properties:
                payload:
                  $ref: '#/definitions/view.UserEmptyView'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/view.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/view.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/view.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/view.Response'
      security:
      - JWT: []
      tags:
      - auth
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"